// frontend/src/interfaces/IItemUse.ts
export interface ItemUsesInterface {
    ID?: number;
    Item_Use_AmountOne?: number;
    Item_Use_ReturnOne?: number;
    Item_Use_AmountTwo?: number;
    Item_Use_ReturnTwo?: number;
    Item_Use_AmountThree?: number;
    Item_Use_ReturnThree?: number;
    Borrowed_Date?: string;
	Due_Date?: string;
	Return_Date?: string;
}